TRUNK-6192: NameSupport.java to provide for NameTemplate overriding via a GP
