+++ api/src/main/java/org/openmrs/layout/name/NameSupport.java
@@ -9,10 +9,18 @@
  */
 package org.openmrs.layout.name;
 
+import javax.xml.parsers.DocumentBuilder;
+import javax.xml.parsers.DocumentBuilderFactory;
+import javax.xml.parsers.ParserConfigurationException;
+import java.io.ByteArrayInputStream;
+import java.io.IOException;
+import java.nio.charset.StandardCharsets;
 import java.util.ArrayList;
 import java.util.List;
 
+import org.apache.commons.lang3.StringUtils;
 import org.openmrs.GlobalProperty;
+import org.openmrs.api.APIException;
 import org.openmrs.api.GlobalPropertyListener;
 import org.openmrs.api.context.Context;
 import org.openmrs.layout.LayoutSupport;
@@ -20,6 +28,8 @@ import org.openmrs.serialization.SerializationException;
 import org.openmrs.util.OpenmrsConstants;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.w3c.dom.Document;
+import org.xml.sax.SAXException;
 
 /**
  * @since 1.12
@@ -29,6 +39,7 @@ public class NameSupport extends LayoutSupport<NameTemplate> implements GlobalPr
 	private static final Logger log = LoggerFactory.getLogger(NameSupport.class);
 	private static NameSupport singleton;
 	private boolean initialized = false;
+	private boolean customTemplateConfigured = false;
 	
 	public NameSupport() {
 		if (singleton == null) {
@@ -38,34 +49,51 @@ public class NameSupport extends LayoutSupport<NameTemplate> implements GlobalPr
 	
 	public static NameSupport getInstance() {
 		synchronized (NameSupport.class) {
-			if (singleton == null) {
+			singleton.init();
+			if (singleton.customTemplateConfigured) {
+				return singleton;
+			} else {
 				new NameSupport();
 			}
+			
+			if (singleton == null) {
+				throw new APIException("Not Yet Instantiated");
+			} else {
+				return singleton;
+			}
 		}
-		singleton.init();
-		return singleton;
 	}
 
 	private void init() {
 		if (!initialized) {
+			Context.getAdministrationService().addGlobalPropertyListener(singleton);
 			try {
-				Context.getAdministrationService().addGlobalPropertyListener(singleton);
 				// Get layout NameTemplate format to override the default if any
 				String layoutTemplateXml = Context.getAdministrationService().getGlobalProperty(
-					OpenmrsConstants.GLOBAL_PROPERTY_LAYOUT_NAME_FORMAT);
-				setNameTemplate(layoutTemplateXml);
+					OpenmrsConstants.GLOBAL_PROPERTY_LAYOUT_NAME_TEMPLATE);
+				
+				if (StringUtils.isNotBlank(layoutTemplateXml)) {
+					// Validating... Custom NameTemplate should be in a Valid XML format
+					DocumentBuilderFactory dbf = DocumentBuilderFactory.newInstance();
+					DocumentBuilder builder = dbf.newDocumentBuilder();
+					Document doc = builder.parse(new ByteArrayInputStream(layoutTemplateXml.getBytes(StandardCharsets.UTF_8)));
+					
+					setNameTemplate(layoutTemplateXml);
+					List<String> specialTokens = new ArrayList<>();
+					specialTokens.add("prefix");
+					specialTokens.add("givenName");
+					specialTokens.add("middleName");
+					specialTokens.add("familyNamePrefix");
+					specialTokens.add("familyName");
+					specialTokens.add("familyName2");
+					specialTokens.add("familyNameSuffix");
+					specialTokens.add("degree");
+					setSpecialTokens(specialTokens);
+					customTemplateConfigured = true;
+				}
+			} catch (ParserConfigurationException | IOException | SAXException e) {
+				log.error("Custom NameTemplate should be in a Valid XML format", e);
 			} catch (NullPointerException ignored) {}
-
-			List<String> specialTokens = new ArrayList<>();
-			specialTokens.add("prefix");
-			specialTokens.add("givenName");
-			specialTokens.add("middleName");
-			specialTokens.add("familyNamePrefix");
-			specialTokens.add("familyName");
-			specialTokens.add("familyName2");
-			specialTokens.add("familyNameSuffix");
-			specialTokens.add("degree");
-			setSpecialTokens(specialTokens);
 			initialized = true;
 		}
 	}
@@ -83,13 +111,7 @@ public class NameSupport extends LayoutSupport<NameTemplate> implements GlobalPr
 			nameTemplate = Context.getSerializationService().getDefaultSerializer().deserialize(xml,
 				NameTemplate.class);
 		} catch (SerializationException | NullPointerException e) {
-			log.error("Error in deserializing provided name template, loading default template", e);
-			try {
-				nameTemplate = Context.getSerializationService().getDefaultSerializer().deserialize(OpenmrsConstants.LONG_NAME_TEMPLATE,
-					NameTemplate.class);
-			} catch (SerializationException | NullPointerException e2) {
-				return;
-			}
+			log.error("Error in deserializing provided name template", e);
 		}
 		
 		List<NameTemplate> list = new ArrayList<NameTemplate>();
@@ -118,8 +140,15 @@ public class NameSupport extends LayoutSupport<NameTemplate> implements GlobalPr
 	 */
 	@Override
 	public String getDefaultLayoutFormat() {
-		String xml = Context.getAdministrationService().getGlobalProperty("layout.name.format");
-		setNameTemplate(xml);
+		try {
+			String xml;
+			xml = Context.getAdministrationService().getGlobalProperty("layout.name.template");
+			if (StringUtils.isBlank(xml)) {
+				xml = Context.getAdministrationService().getGlobalProperty("layout.name.format");
+				return xml != null && xml.length() > 0 ? xml : this.defaultLayoutFormat;
+			}
+			setNameTemplate(xml);
+		} catch (NullPointerException ignored) {}
 		return (this.layoutTemplates != null && this.layoutTemplates.size() > 0) ? this.layoutTemplates.get(0).getCodeName() : defaultLayoutFormat;
 	}
